refactor: wrap event dialogs in ModalBottomSheet and rename App to CalendarApp
This commit refactors the `AddEventDialog` and `EventDetailsDialog` to be displayed within `ModalBottomSheet` components. Additionally, the main application composable `App` has been renamed to `CalendarApp` for clarity.

**Key Changes:**

-   **`AddEventDialog` & `EventDetailsDialog`:**
    -   Both dialogs are now wrapped in `ModalBottomSheet`.
    -   A `sheetState` using `rememberModalBottomSheetState(skipPartiallyExpanded = true)` is initialized for each.
    -   `ModalBottomSheetProperties` are configured to `shouldDismissOnBackPress = true`.
    -   The internal `Column` layout and logic of the dialogs remain largely the same, but are now hosted within the bottom sheet.
-   **`App.kt` renamed to `CalendarApp.kt`:**
    -   The main composable function `App()` is renamed to `CalendarApp()`.
    -   All usages of `App()` (in `MainActivity.kt`, `Main.kt` for desktop, and `MainViewController.kt` for iOS) have been updated to `CalendarApp()`.
-   **BottomSheet Invocation in `CalendarApp.kt`:**
    -   The existing `ModalBottomSheet` calls within `CalendarApp` for showing `AddEventDialog` and `EventDetailsDialog` were removed as the dialogs themselves now manage their bottom sheet presentation. The `showAddBottomSheet` and `showDetailsBottomSheet` state variables still control their visibility.
-   **Minor Import Change:**
    -   Removed `ExperimentalSharedTransitionApi` import from `CalendarApp.kt` as it was no longer used.

This change aims to provide a more consistent and modern user experience by utilizing bottom sheets for event creation and detail viewing, aligning with common mobile UI patterns. The renaming of `App` to `CalendarApp` improves code readability and descriptiveness.
