+++ composeApp/src/androidMain/kotlin/com/debanshu/xcalendar/MainActivity.kt
@@ -33,7 +33,7 @@ class MainActivity : ComponentActivity() {
         isDarkMode = resources.configuration.isNightModeActive
         enableEdgeToEdge(transparentBarStyle, transparentBarStyle)
         setContent {
-            App()
+            CalendarApp()
         }
     }
 }
