+++ composeApp/src/desktopMain/kotlin/com/debanshu/xcalendar/Main.kt
@@ -11,7 +11,7 @@ fun main() {
             onCloseRequest = ::exitApplication,
             title = "XCalendar",
         ) {
-            App()
+            CalendarApp()
         }
     }
 }
