+++ composeApp/src/iosMain/kotlin/com/debanshu/xcalendar/MainViewController.kt
@@ -3,7 +3,8 @@ package com.debanshu.xcalendar
 import androidx.compose.ui.window.ComposeUIViewController
 import com.debanshu.xcalendar.di.initKoin
 
-fun MainViewController() = ComposeUIViewController {
-    initKoin()
-    App()
-}
+fun MainViewController() =
+    ComposeUIViewController {
+        initKoin()
+        CalendarApp()
+    }
